Add sum mo stuff
This is the start of adding the "standard" API calls
to the shell as well as to the client.

Note this also includes a fix to make the debug option work.
So now if you specify --debug on the command line you'll get
the raw data for the API call as well as the raw response data.

Adds the following methods:
* GetVolumeStats
* GetDefaultQoS
* ListActiveVolumes
* ListVolumesForAccount