Add unit test framework and tox.ini
Just add the framework for tetss and a dummy placeholder test.
Using unittest in conjunction with tox.

Also adding in pep8 checks and fixing pep8 errors that currently
exist.

To run tests for nox, just use "'tox', 'tox -epy27' or 'tox -epep8'"
we'll create a pretty wrapper later as well as some proper tests
and docs.