Update on API strategy
This updates to use all of the auto-generated python bindings with
some modifications.  We take the auto-generated code and modify it
to fit proper python naming and formatting.  We also do some things
to extract it more appropriately and dump the reference/response
object passing.  It's simple enough to just pass the results into
a simple helper method and conver it to an object automatically.

Note we also get rid of the "double implementation" thing we had
going and just implement the single pythonic "list_volumes" instead
of "list_voumes" and "ListVolumes".

There's still a bit of work to update the usage in shell and client
and we need to get unit tests worked, but we now have the framework
in place so it should be relatively easy to roll.