Rework using Click and focus on base methods first
This branch takes a step back, and focuses mostly on
just using what's available via our modified version of
jsvcgen.

We create a solidfire_element_api module with the Cluster's
base methods as well as a generic request method.  For now
we just build up the CLI using that module.

Going forward, we also have the managers directory where more
complex tasks that chain methods together should live.  The
idea here is that at some point we implement helpers for things
like backups and replication etc.

Additionally, we also add the ability for consumers to add a
path to their own modules as well and automatically import those.
This way they can leverage the base work as well as have their own
custom/secret methods to use as well.